requirements.txt
absl-py==2.0.0
altair==5.1.2
astunparse==1.6.3
attrs==23.1.0
black==23.11.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
flatbuffers==23.5.26
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
jax==0.4.20
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
keras==2.12.0
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.3.1
mypy-extensions==1.0.0
numpy==1.23.5
nvidia-cudnn==8.2.0.51
nvidia-pyindex==1.0.9
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.0.0
progressbar2==4.2.0
protobuf==4.25.1
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.17.1
python-barcode==0.15.1
python-dateutil==2.8.2
python-utils==3.8.1
pytz==2023.3.post1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
scipy==1.11.4
six==1.16.0
smmap==5.0.1
streamlit==1.28.2
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
